Fixed: Form redirecting back to form after submitting input
Submitting form input does not make the thank you page appear because  the textbox doesn't have
a name attribute to it. Because of this, checking if an array key for the name exists in $_GET
always result to False, setting the template that will be used next to be the form page.
By giving a name to the textbox, the problem is fixed.

> textbox name is 'name'
> $_GET key is now 'name'